requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
